+++ apps/backend/src/deployer.py
@@ -1,39 +1,26 @@
-from typing import Any, Callable, Final
+from typing import Any, Callable
 from fastapi import FastAPI
 
-import re
-import os
-import importlib
+import routes.auth.get as getAuth
+import routes.auth.post as postAuth
+import routes.auth.put as putAuth
 
-ROUTES_DIR: Final[str] = "routes"
+import routes.rooms.roomId_.my_socket as my_socketRoomsRoomId_
 
-ROUTES = list(filter(lambda x: x.endswith(".py"), os.listdir(ROUTES_DIR)))
-ROUTE_REX_EX = r"(?:.*_?)*__.*"
+import routes.user.get as getUser
 
-def main() -> dict[str, Callable[[Any, Any], Any]]:
-    modules: dict[str, Callable[[Any, Any], Any]] = {}
-
-    for route in ROUTES:
-        data, _ = route.split(".")
-        
-        if re.match(ROUTE_REX_EX, data) == None: continue
-
-        path, method = data.split("__")
-
-        module: Any = importlib.import_module(f"{ROUTES_DIR}.{path}__{method}")
-        modules[f"{path.replace("_", "/")}.{method}"] = module.main
-
-    return modules
+routes: list[Callable[[FastAPI], Any]] = [
+    getAuth.main,
+    postAuth.main,
+    putAuth.main,
+    my_socketRoomsRoomId_.main,
+    getUser.main,
+]
 
 def register(app: FastAPI):
-    routes = main()
-
     for route in routes:
-        routes[route](app, routes)
+        route(app)
 
 if __name__ == "__main__":
-    routes = main()
-
     for route in routes:
         print(route)
-        print(routes[route])
