+++ apps/backend/src/routes/auth/put.py
@@ -4,11 +4,11 @@ from fastapi.responses import JSONResponse
 from database.schemas import AuthBase
 from database.session import Database
 
-from .tokens import validate, fetch, generateToken
+from ..tokens import validate, fetch, generateToken
 
 import hash
 
-def main(app: FastAPI, _):
+def main(app: FastAPI):
     @app.put("/api/auth")
     async def execute(req: Request):
         token = req.headers.get("authorization")
